:fix 키워드 업데이트 로직 최적화
