+++ src/main/java/com/jms/alertmessaging/service/crawl/v1/CrawlingService.java
@@ -1,4 +1,4 @@
-package com.jms.alertmessaging.service.crawl;
+package com.jms.alertmessaging.service.crawl.v1;
 
 import com.jms.alertmessaging.entity.board.Board;
 import com.jms.alertmessaging.entity.department.Department;
